id_service/requirements.txt
# id_service/requirements.txt

# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis==5.2.0

# Security & Auth
authlib==1.3.2
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
cryptography==43.0.3

# HTTP & Forms
python-multipart==0.0.12
httpx==0.27.2

# Email
aiosmtplib==3.0.2
email-validator==2.2.0
jinja2==3.1.4

# Utilities
python-dotenv==1.0.1
itsdangerous==2.2.0

# Development (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.4